[OPS-145 / #40]refactor: 모든 api ResponseEntity 반환하도록 리팩토링
